+++ aten/src/ATen/native/quantized/cpu/qhardsigmoid.cpp
@@ -43,7 +43,7 @@ Tensor qnnpack_hardsigmoid(Tensor input) {
   const pytorch_qnnp_status createStatus = pytorch_qnnp_create_hardsigmoid_nc_q8(
     num_elems, // channels
     i_zero_point,
-    i_scale,
+    static_cast<float>(i_scale),
     o_zero_point,
     o_scale,
     std::numeric_limits<uint8_t>::min(), // output min
